sql/randgen: refactor randgen package
This commit splits the single file of the `randgen` package into
multiple files organized by the theme of the contained utility
functions.

This commit also moves `randgen.OneIntCol`, `randgen.TwoIntCols`, and
other non-random, type-related test helpers to the `types` package in
`pkg/sql/types/testutils.go`.

Release note: None
