+++ pkg/sql/colexec/default_agg_test.go
@@ -21,7 +21,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colexec/colexectestutils"
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecop"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
-	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
@@ -110,7 +109,7 @@ func TestDefaultAggregateFunc(t *testing.T) {
 		},
 		{
 			name: "XorAgg",
-			typs: rowenc.TwoIntCols,
+			typs: types.TwoIntCols,
 			input: colexectestutils.Tuples{
 				{nil, 3},
 				{nil, 1},
