+++ pkg/sql/colexec/hash_aggregator_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecop"
 	"github.com/cockroachdb/cockroach/pkg/sql/colmem"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
-	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/testutils/colcontainerutils"
@@ -37,7 +36,7 @@ import (
 var hashAggregatorTestCases = []aggregatorTestCase{
 	{
 		name: "carryBetweenBatches",
-		typs: rowenc.TwoIntCols,
+		typs: types.TwoIntCols,
 		input: colexectestutils.Tuples{
 			{0, 1},
 			{1, 5},
@@ -62,7 +61,7 @@ var hashAggregatorTestCases = []aggregatorTestCase{
 	},
 	{
 		name: "bucketCollision",
-		typs: rowenc.TwoIntCols,
+		typs: types.TwoIntCols,
 		input: colexectestutils.Tuples{
 			{0, 3},
 			{0, 4},
