+++ pkg/sql/colflow/colbatch_scan_test.go
@@ -29,8 +29,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colfetcher"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
-	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
@@ -82,7 +82,7 @@ func TestColBatchScanMeta(t *testing.T) {
 			Projection:    true,
 			OutputColumns: []uint32{0},
 		},
-		ResultTypes: randgen.OneIntCol,
+		ResultTypes: types.OneIntCol,
 	}
 
 	args := &colexecargs.NewColOperatorArgs{
@@ -140,7 +140,7 @@ func BenchmarkColBatchScan(b *testing.B) {
 					Projection:    true,
 					OutputColumns: []uint32{0, 1},
 				},
-				ResultTypes: randgen.TwoIntCols,
+				ResultTypes: types.TwoIntCols,
 			}
 
 			evalCtx := tree.MakeTestingEvalContext(s.ClusterSettings())
