+++ pkg/sql/colflow/vectorized_panic_propagation_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/errors"
 	"github.com/stretchr/testify/require"
@@ -45,8 +46,8 @@ func TestVectorizedInternalPanic(t *testing.T) {
 	}
 
 	nRows, nCols := 1, 1
-	types := randgen.OneIntCol
-	input := execinfra.NewRepeatableRowSource(types, randgen.MakeIntRows(nRows, nCols))
+	typs := types.OneIntCol
+	input := execinfra.NewRepeatableRowSource(typs, randgen.MakeIntRows(nRows, nCols))
 
 	col, err := colexec.NewBufferingColumnarizer(ctx, testAllocator, &flowCtx, 0 /* processorID */, input)
 	if err != nil {
@@ -58,7 +59,7 @@ func TestVectorizedInternalPanic(t *testing.T) {
 		&flowCtx,
 		1, /* processorID */
 		vee,
-		types,
+		typs,
 		nil, /* output */
 		nil, /* getStats */
 		nil, /* metadataSourceQueue */
@@ -92,8 +93,8 @@ func TestNonVectorizedPanicPropagation(t *testing.T) {
 	}
 
 	nRows, nCols := 1, 1
-	types := randgen.OneIntCol
-	input := execinfra.NewRepeatableRowSource(types, randgen.MakeIntRows(nRows, nCols))
+	typs := types.OneIntCol
+	input := execinfra.NewRepeatableRowSource(typs, randgen.MakeIntRows(nRows, nCols))
 
 	col, err := colexec.NewBufferingColumnarizer(ctx, testAllocator, &flowCtx, 0 /* processorID */, input)
 	if err != nil {
@@ -105,7 +106,7 @@ func TestNonVectorizedPanicPropagation(t *testing.T) {
 		&flowCtx,
 		1, /* processorID */
 		nvee,
-		types,
+		typs,
 		nil, /* output */
 		nil, /* getStats */
 		nil, /* metadataSourceQueue */
