+++ pkg/sql/flowinfra/outbox_test.go
@@ -27,7 +27,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/flowinfra"
-	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
@@ -80,7 +79,7 @@ func TestOutbox(t *testing.T) {
 	}
 	streamID := execinfrapb.StreamID(42)
 	outbox := flowinfra.NewOutbox(&flowCtx, execinfra.StaticNodeID, streamID, nil /* numOutboxes */, false /* isGatewayNode */)
-	outbox.Init(randgen.OneIntCol)
+	outbox.Init(types.OneIntCol)
 	var outboxWG sync.WaitGroup
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -156,12 +155,12 @@ func TestOutbox(t *testing.T) {
 		// about the draining.
 		last := -1
 		for i := 0; i < len(rows); i++ {
-			if rows[i].String(randgen.OneIntCol) != "[-1]" {
+			if rows[i].String(types.OneIntCol) != "[-1]" {
 				last = i
 				continue
 			}
 			for j := i; j < len(rows); j++ {
-				if rows[j].String(randgen.OneIntCol) == "[-1]" {
+				if rows[j].String(types.OneIntCol) == "[-1]" {
 					continue
 				}
 				rows[i] = rows[j]
@@ -194,7 +193,7 @@ func TestOutbox(t *testing.T) {
 			t.Fatalf("expected: %q, got: %q", expectedStr, m.Err.Error())
 		}
 	}
-	str := rows.String(randgen.OneIntCol)
+	str := rows.String(types.OneIntCol)
 	expected := "[[0]]"
 	if str != expected {
 		t.Errorf("invalid results: %s, expected %s'", str, expected)
@@ -242,7 +241,7 @@ func TestOutboxInitializesStreamBeforeReceivingAnyRows(t *testing.T) {
 	var outboxWG sync.WaitGroup
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	outbox.Init(randgen.OneIntCol)
+	outbox.Init(types.OneIntCol)
 	// Start the outbox. This should cause the stream to connect, even though
 	// we're not sending any rows.
 	outbox.Start(ctx, &outboxWG, cancel)
@@ -315,7 +314,7 @@ func TestOutboxClosesWhenConsumerCloses(t *testing.T) {
 			ctx, cancel := context.WithCancel(context.Background())
 			defer cancel()
 			outbox = flowinfra.NewOutbox(&flowCtx, execinfra.StaticNodeID, streamID, nil /* numOutboxes */, false /* isGatewayNode */)
-			outbox.Init(randgen.OneIntCol)
+			outbox.Init(types.OneIntCol)
 			outbox.Start(ctx, &wg, cancel)
 
 			// Wait for the outbox to connect the stream.
@@ -395,7 +394,7 @@ func TestOutboxCancelsFlowOnError(t *testing.T) {
 	}
 
 	outbox = flowinfra.NewOutbox(&flowCtx, execinfra.StaticNodeID, streamID, nil /* numOutboxes */, false /* isGatewayNode */)
-	outbox.Init(randgen.OneIntCol)
+	outbox.Init(types.OneIntCol)
 	outbox.Start(ctx, &wg, mockCancel)
 
 	// Wait for the outbox to connect the stream.
@@ -443,7 +442,7 @@ func TestOutboxUnblocksProducers(t *testing.T) {
 	defer cancel()
 
 	outbox = flowinfra.NewOutbox(&flowCtx, execinfra.StaticNodeID, streamID, nil /* numOutboxes */, false /* isGatewayNode */)
-	outbox.Init(randgen.OneIntCol)
+	outbox.Init(types.OneIntCol)
 
 	// Fill up the outbox.
 	for i := 0; i < flowinfra.OutboxBufRows; i++ {
@@ -509,7 +508,7 @@ func BenchmarkOutbox(b *testing.B) {
 				NodeID: base.TestingIDContainer,
 			}
 			outbox := flowinfra.NewOutbox(&flowCtx, execinfra.StaticNodeID, streamID, nil /* numOutboxes */, false /* isGatewayNode */)
-			outbox.Init(randgen.MakeIntCols(numCols))
+			outbox.Init(types.MakeIntCols(numCols))
 			var outboxWG sync.WaitGroup
 			ctx, cancel := context.WithCancel(context.Background())
 			defer cancel()
