+++ pkg/sql/rowexec/aggregator_test.go
@@ -74,7 +74,7 @@ func TestAggregator(t *testing.T) {
 			Name: "MinMaxCountAvgSumStddevGroupByNoneNoRows",
 			Input: ProcessorTestCaseRows{
 				Rows:  [][]interface{}{},
-				Types: randgen.MakeIntCols(1),
+				Types: types.MakeIntCols(1),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
@@ -107,7 +107,7 @@ func TestAggregator(t *testing.T) {
 					{7, 2},
 					{8, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
@@ -115,7 +115,7 @@ func TestAggregator(t *testing.T) {
 					{4, 1},
 					{2, 3},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -138,14 +138,14 @@ func TestAggregator(t *testing.T) {
 					{7, 2},
 					{8, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
 					{4, 2},
 					{2, 3},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -168,14 +168,14 @@ func TestAggregator(t *testing.T) {
 					{3, 4},
 					{8, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
 					{2, 3},
 					{4, 2},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			DisableSort: true,
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
@@ -200,7 +200,7 @@ func TestAggregator(t *testing.T) {
 					{7, 2},
 					{8, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
@@ -230,7 +230,7 @@ func TestAggregator(t *testing.T) {
 					{8, 4},
 					{3, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
@@ -262,7 +262,7 @@ func TestAggregator(t *testing.T) {
 					{4, 2},
 					{5, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
@@ -290,13 +290,13 @@ func TestAggregator(t *testing.T) {
 					{2},
 					{4},
 				},
-				Types: randgen.MakeIntCols(1),
+				Types: types.MakeIntCols(1),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
 					{6},
 				},
-				Types: randgen.MakeIntCols(1),
+				Types: types.MakeIntCols(1),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -315,13 +315,13 @@ func TestAggregator(t *testing.T) {
 					{1},
 					{1},
 				},
-				Types: randgen.MakeIntCols(1),
+				Types: types.MakeIntCols(1),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
 					{1},
 				},
-				Types: randgen.MakeIntCols(1),
+				Types: types.MakeIntCols(1),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -341,13 +341,13 @@ func TestAggregator(t *testing.T) {
 					{4, 2},
 					{5, 4},
 				},
-				Types: randgen.MakeIntCols(2),
+				Types: types.MakeIntCols(2),
 			},
 			Output: ProcessorTestCaseRows{
 				Rows: [][]interface{}{
 					{5, 2, 5, 2},
 				},
-				Types: randgen.MakeIntCols(4),
+				Types: types.MakeIntCols(4),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -376,7 +376,7 @@ func TestAggregator(t *testing.T) {
 				Rows: [][]interface{}{
 					{2, 3, 3},
 				},
-				Types: randgen.MakeIntCols(3),
+				Types: types.MakeIntCols(3),
 			},
 			ProcessorCore: execinfrapb.ProcessorCoreUnion{
 				Aggregator: &execinfrapb.AggregatorSpec{
@@ -436,7 +436,7 @@ func BenchmarkAggregation(b *testing.B) {
 			}
 			post := &execinfrapb.PostProcessSpec{}
 			disposer := &rowDisposer{}
-			input := execinfra.NewRepeatableRowSource(randgen.OneIntCol, randgen.MakeIntRows(numRows, numCols))
+			input := execinfra.NewRepeatableRowSource(types.OneIntCol, randgen.MakeIntRows(numRows, numCols))
 
 			b.SetBytes(int64(8 * numRows * numCols))
 			b.ResetTimer()
@@ -466,7 +466,7 @@ func BenchmarkCountRows(b *testing.B) {
 	disposer := &rowDisposer{}
 	const numCols = 1
 	const numRows = 100000
-	input := execinfra.NewRepeatableRowSource(randgen.OneIntCol, randgen.MakeIntRows(numRows, numCols))
+	input := execinfra.NewRepeatableRowSource(types.OneIntCol, randgen.MakeIntRows(numRows, numCols))
 
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
@@ -509,7 +509,7 @@ func BenchmarkGrouping(b *testing.B) {
 	}
 	post := &execinfrapb.PostProcessSpec{}
 	disposer := &rowDisposer{}
-	input := execinfra.NewRepeatableRowSource(randgen.OneIntCol, randgen.MakeIntRows(numRows, numCols))
+	input := execinfra.NewRepeatableRowSource(types.OneIntCol, randgen.MakeIntRows(numRows, numCols))
 
 	b.SetBytes(int64(8 * numRows * numCols))
 	b.ResetTimer()
@@ -568,7 +568,7 @@ func benchmarkAggregationWithGrouping(b *testing.B, numOrderedCols int) {
 			spec.OrderedGroupCols = allOrderedGroupCols[:numOrderedCols]
 			post := &execinfrapb.PostProcessSpec{}
 			disposer := &rowDisposer{}
-			input := execinfra.NewRepeatableRowSource(randgen.ThreeIntCols, makeGroupedIntRows(groupSize, numCols, groupedCols[:]))
+			input := execinfra.NewRepeatableRowSource(types.ThreeIntCols, makeGroupedIntRows(groupSize, numCols, groupedCols[:]))
 
 			b.SetBytes(int64(8 * intPow(groupSize, len(groupedCols)+1) * numCols))
 			b.ResetTimer()
