+++ pkg/sql/rowexec/distinct_test.go
@@ -268,7 +268,7 @@ func TestDistinct(t *testing.T) {
 		t.Run("", func(t *testing.T) {
 			ds := c.spec
 
-			in := distsqlutils.NewRowBuffer(randgen.ThreeIntCols, c.input, distsqlutils.RowBufferArgs{})
+			in := distsqlutils.NewRowBuffer(types.ThreeIntCols, c.input, distsqlutils.RowBufferArgs{})
 			out := &distsqlutils.RowBuffer{}
 
 			st := cluster.MakeTestingClusterSettings()
@@ -306,8 +306,8 @@ func TestDistinct(t *testing.T) {
 					t.Errorf("expected error: %v, got %v", c.error, err)
 				}
 			} else {
-				if result := res.String(randgen.ThreeIntCols); result != c.expected.String(randgen.ThreeIntCols) {
-					t.Errorf("invalid results: %v, expected %v'", result, c.expected.String(randgen.ThreeIntCols))
+				if result := res.String(types.ThreeIntCols); result != c.expected.String(types.ThreeIntCols) {
+					t.Errorf("invalid results: %v, expected %v'", result, c.expected.String(types.ThreeIntCols))
 				}
 			}
 		})
@@ -335,7 +335,7 @@ func benchmarkDistinct(b *testing.B, orderedColumns []uint32) {
 	post := &execinfrapb.PostProcessSpec{}
 	for _, numRows := range []int{1 << 4, 1 << 8, 1 << 12, 1 << 16} {
 		b.Run(fmt.Sprintf("rows=%d", numRows), func(b *testing.B) {
-			input := execinfra.NewRepeatableRowSource(randgen.TwoIntCols, randgen.MakeIntRows(numRows, numCols))
+			input := execinfra.NewRepeatableRowSource(types.TwoIntCols, randgen.MakeIntRows(numRows, numCols))
 
 			b.SetBytes(int64(8 * numRows * numCols))
 			b.ResetTimer()
