+++ pkg/sql/rowexec/inverted_joiner_test.go
@@ -25,7 +25,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/inverted"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/invertedexpr"
-	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
@@ -280,10 +279,10 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes:               randgen.OneIntCol,
+			inputTypes:               types.OneIntCol,
 			datumsToExpr:             arrayIntersectionExpr{t: t},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.TwoIntCols,
+			outputTypes:              types.TwoIntCols,
 			outputColumns:            []uint32{0, 1},
 			expected:                 "[[5 5] [5 50] [20 2] [20 20] [42 24] [42 42]]",
 			expectedWithContinuation: "[[5 5 false] [5 50 true] [20 2 false] [20 20 true] [42 24 false] [42 42 true]]",
@@ -298,12 +297,12 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes: randgen.OneIntCol,
+			inputTypes: types.OneIntCol,
 			datumsToExpr: arrayIntersectionExpr{
 				t: t, toExclude: &struct{ left, right int64 }{left: 5, right: 0},
 			},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.TwoIntCols,
+			outputTypes:              types.TwoIntCols,
 			outputColumns:            []uint32{0, 1},
 			expected:                 "[[20 2] [20 20] [42 24] [42 42]]",
 			expectedWithContinuation: "[[20 2 false] [20 20 true] [42 24 false] [42 42 true]]",
@@ -320,10 +319,10 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes:               randgen.OneIntCol,
+			inputTypes:               types.OneIntCol,
 			datumsToExpr:             arrayIntersectionExpr{t: t},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.TwoIntCols,
+			outputTypes:              types.TwoIntCols,
 			outputColumns:            []uint32{0, 1},
 			expected:                 "[[5 50] [42 24] [42 42]]",
 			expectedWithContinuation: "[[5 50 false] [42 24 false] [42 42 true]]",
@@ -338,10 +337,10 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  arrayIntersectionExpr{t: t},
 			joinType:      descpb.LeftOuterJoin,
-			outputTypes:   randgen.TwoIntCols,
+			outputTypes:   types.TwoIntCols,
 			outputColumns: []uint32{0, 1},
 			// Similar to previous, but the left side input failing the onExpr is emitted.
 			expected:                 "[[5 50] [20 NULL] [42 24] [42 42]]",
@@ -356,10 +355,10 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  arrayIntersectionExpr{t: t},
 			joinType:      descpb.LeftSemiJoin,
-			outputTypes:   randgen.OneIntCol,
+			outputTypes:   types.OneIntCol,
 			outputColumns: []uint32{0},
 			expected:      "[[5] [42]]",
 		},
@@ -372,10 +371,10 @@ func TestInvertedJoiner(t *testing.T) {
 			input: [][]tree.Datum{
 				{tree.NewDInt(tree.DInt(5))}, {tree.NewDInt(tree.DInt(20))}, {tree.NewDInt(tree.DInt(42))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  arrayIntersectionExpr{t: t},
 			joinType:      descpb.LeftAntiJoin,
-			outputTypes:   randgen.OneIntCol,
+			outputTypes:   types.OneIntCol,
 			outputColumns: []uint32{0},
 			expected:      "[[20]]",
 		},
@@ -391,10 +390,10 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101))},
 				{tree.NewDInt(tree.DInt(20))},
 			},
-			inputTypes:               randgen.OneIntCol,
+			inputTypes:               types.OneIntCol,
 			datumsToExpr:             jsonIntersectionExpr{},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.TwoIntCols,
+			outputTypes:              types.TwoIntCols,
 			outputColumns:            []uint32{0, 1},
 			expected:                 "[[5 5] [42 42] [20 20]]",
 			expectedWithContinuation: "[[5 5 false] [42 42 false] [20 20 false]]",
@@ -412,10 +411,10 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101))},
 				{tree.NewDInt(tree.DInt(20))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  jsonUnionExpr{},
 			joinType:      descpb.InnerJoin,
-			outputTypes:   randgen.TwoIntCols,
+			outputTypes:   types.TwoIntCols,
 			outputColumns: []uint32{0, 1},
 			// The ordering looks odd because of how the JSON values were ordered in the
 			// inverted index. The spans we are reading for the first batch of two inputs
@@ -455,10 +454,10 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101))},
 				{tree.NewDInt(tree.DInt(20))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  jsonUnionExpr{},
 			joinType:      descpb.LeftSemiJoin,
-			outputTypes:   randgen.OneIntCol,
+			outputTypes:   types.OneIntCol,
 			outputColumns: []uint32{0},
 			expected:      "[[5] [42] [101] [20]]",
 		},
@@ -472,10 +471,10 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101))},
 				{tree.NewDInt(tree.DInt(20))},
 			},
-			inputTypes:    randgen.OneIntCol,
+			inputTypes:    types.OneIntCol,
 			datumsToExpr:  jsonUnionExpr{},
 			joinType:      descpb.LeftAntiJoin,
-			outputTypes:   randgen.OneIntCol,
+			outputTypes:   types.OneIntCol,
 			outputColumns: []uint32{0},
 			expected:      "[]",
 		},
@@ -494,11 +493,11 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(20)), tree.NewDInt(tree.DInt(0))},
 				{tree.NewDInt(tree.DInt(42)), tree.NewDInt(tree.DInt(420))},
 			},
-			inputTypes:               randgen.TwoIntCols,
+			inputTypes:               types.TwoIntCols,
 			prefixEqualityColumns:    []uint32{1},
 			datumsToExpr:             arrayIntersectionExpr{t: t},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[5 50 5] [42 420 42]]",
 			expectedWithContinuation: "[[5 50 5 false] [42 420 42 false]]",
@@ -514,13 +513,13 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(21)), tree.NewDInt(tree.DInt(0))},
 				{tree.NewDInt(tree.DInt(42)), tree.NewDInt(tree.DInt(420))},
 			},
-			inputTypes:            randgen.TwoIntCols,
+			inputTypes:            types.TwoIntCols,
 			prefixEqualityColumns: []uint32{1},
 			datumsToExpr: arrayIntersectionExpr{
 				t: t, toExclude: &struct{ left, right int64 }{left: 5, right: 0},
 			},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[20 200 20] [42 420 42]]",
 			expectedWithContinuation: "[[20 200 20 false] [42 420 42 false]]",
@@ -539,11 +538,11 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(21)), tree.NewDInt(tree.DInt(0))},
 				{tree.NewDInt(tree.DInt(42)), tree.NewDInt(tree.DInt(420))},
 			},
-			inputTypes:               randgen.TwoIntCols,
+			inputTypes:               types.TwoIntCols,
 			prefixEqualityColumns:    []uint32{1},
 			datumsToExpr:             arrayIntersectionExpr{t: t},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[42 420 42]]",
 			expectedWithContinuation: "[[42 420 42 false]]",
@@ -562,11 +561,11 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101)), tree.NewDInt(tree.DInt(1010))},
 				{tree.NewDInt(tree.DInt(20)), tree.NewDInt(tree.DInt(200))},
 			},
-			inputTypes:               randgen.TwoIntCols,
+			inputTypes:               types.TwoIntCols,
 			prefixEqualityColumns:    []uint32{1},
 			datumsToExpr:             jsonIntersectionExpr{},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[5 50 5] [20 200 20]]",
 			expectedWithContinuation: "[[5 50 5 false] [20 200 20 false]]",
@@ -584,11 +583,11 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101)), tree.NewDInt(tree.DInt(1010))},
 				{tree.NewDInt(tree.DInt(20)), tree.NewDInt(tree.DInt(200))},
 			},
-			inputTypes:               randgen.TwoIntCols,
+			inputTypes:               types.TwoIntCols,
 			prefixEqualityColumns:    []uint32{1},
 			datumsToExpr:             jsonUnionExpr{},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[5 50 5] [42 420 42] [20 200 20]]",
 			expectedWithContinuation: "[[5 50 5 false] [42 420 42 false] [20 200 20 false]]",
@@ -606,11 +605,11 @@ func TestInvertedJoiner(t *testing.T) {
 				{tree.NewDInt(tree.DInt(101)), tree.NewDInt(tree.DInt(1010))},
 				{tree.NewDInt(tree.DInt(20)), tree.NewDInt(tree.DInt(200))},
 			},
-			inputTypes:               randgen.TwoIntCols,
+			inputTypes:               types.TwoIntCols,
 			prefixEqualityColumns:    []uint32{1, 0},
 			datumsToExpr:             jsonIntersectionExpr{},
 			joinType:                 descpb.InnerJoin,
-			outputTypes:              randgen.ThreeIntCols,
+			outputTypes:              types.ThreeIntCols,
 			outputColumns:            []uint32{0, 1, 2},
 			expected:                 "[[5 50 5] [42 420 42] [20 200 20]]",
 			expectedWithContinuation: "[[5 50 5 false] [42 420 42 false] [20 200 20 false]]",
@@ -776,7 +775,7 @@ func TestInvertedJoinerDrain(t *testing.T) {
 				Type:         descpb.InnerJoin,
 			},
 			arrayIntersectionExpr{t: t},
-			distsqlutils.NewRowBuffer(randgen.TwoIntCols, nil /* rows */, distsqlutils.RowBufferArgs{}),
+			distsqlutils.NewRowBuffer(types.TwoIntCols, nil /* rows */, distsqlutils.RowBufferArgs{}),
 			&execinfrapb.PostProcessSpec{},
 			out,
 		)
