+++ pkg/sql/rowexec/mergejoiner_test.go
@@ -66,7 +66,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 3, 4},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[1], v[4]},
@@ -75,13 +75,13 @@ func TestMergeJoiner(t *testing.T) {
 				{v[4], v[5]},
 				{v[5], v[5]},
 			},
-			rightTypes: randgen.ThreeIntCols,
+			rightTypes: types.ThreeIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
 				{v[4], v[4], v[5]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
@@ -102,12 +102,12 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 1, 3},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[4]},
 				{v[0], v[1]},
@@ -115,7 +115,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[0], v[5]},
 				{v[0], v[4]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0], v[4]},
 				{v[0], v[0], v[1]},
@@ -144,14 +144,14 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit AND @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 1, 3},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
 				{v[1], v[0]},
 				{v[1], v[1]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[4]},
 				{v[0], v[1]},
@@ -164,7 +164,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[1], v[5]},
 				{v[1], v[4]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0], v[4]},
 				{v[0], v[0], v[5]},
@@ -195,7 +195,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit AND @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 1, 3},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[0]},
@@ -209,7 +209,7 @@ func TestMergeJoiner(t *testing.T) {
 
 				{v[6], v[0]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[5]},
 				{v[0], v[5]},
@@ -224,7 +224,7 @@ func TestMergeJoiner(t *testing.T) {
 
 				{v[5], v[0]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0], null},
 				{v[0], v[0], null},
@@ -259,7 +259,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 3, 4},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[1], v[4]},
@@ -268,13 +268,13 @@ func TestMergeJoiner(t *testing.T) {
 				{v[4], v[5]},
 				{v[5], v[5]},
 			},
-			rightTypes: randgen.ThreeIntCols,
+			rightTypes: types.ThreeIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
 				{v[4], v[4], v[5]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], null, null},
 				{v[1], v[0], v[4]},
@@ -298,13 +298,13 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit @1 = @3 constraint.
 			},
 			outCols:   []uint32{3, 1, 2},
-			leftTypes: randgen.ThreeIntCols,
+			leftTypes: types.ThreeIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
 				{v[4], v[4], v[5]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[1], v[4]},
@@ -313,7 +313,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[4], v[5]},
 				{v[5], v[5]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], null, null},
 				{v[1], v[0], v[4]},
@@ -337,7 +337,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 3, 4},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[1], v[4]},
@@ -345,14 +345,14 @@ func TestMergeJoiner(t *testing.T) {
 				{v[3], v[1]},
 				{v[4], v[5]},
 			},
-			rightTypes: randgen.ThreeIntCols,
+			rightTypes: types.ThreeIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
 				{v[4], v[4], v[5]},
 				{v[5], v[5], v[1]},
 			},
-			expectedTypes: randgen.ThreeIntCols,
+			expectedTypes: types.ThreeIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], null, null},
 				{v[1], v[0], v[4]},
@@ -377,14 +377,14 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.FullOuterJoin,
 			},
 			outCols:   []uint32{0, 1, 2, 3},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, v[4]},
 				{v[0], null},
 				{v[0], v[1]},
 				{v[2], v[4]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{null, v[4]},
 				{v[0], null},
@@ -415,17 +415,17 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.InnerJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.OneIntCol,
+			leftTypes: types.OneIntCol,
 			leftInput: rowenc.EncDatumRows{
 				{null},
 				{v[0]},
 			},
-			rightTypes: randgen.OneIntCol,
+			rightTypes: types.OneIntCol,
 			rightInput: rowenc.EncDatumRows{
 				{null},
 				{v[1]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected:      rowenc.EncDatumRows{},
 		},
 		{
@@ -443,18 +443,18 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.LeftSemiJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[1], v[2]},
 				{v[2], v[3]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[2], v[2]},
 				{v[2], v[2]},
 				{v[3], v[3]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[2], v[3]},
 			},
@@ -474,7 +474,7 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.LeftSemiJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[1], v[2]},
 				{v[1], v[2]},
@@ -482,12 +482,12 @@ func TestMergeJoiner(t *testing.T) {
 				{v[3], v[4]},
 				{v[3], v[5]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[2], v[2]},
 				{v[3], v[3]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[2], v[3]},
 				{v[3], v[4]},
@@ -508,18 +508,18 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.LeftSemiJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, v[2]},
 				{v[2], v[3]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{null, v[3]},
 				{v[2], v[4]},
 				{v[2], v[5]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[2], v[3]},
 			},
@@ -540,7 +540,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit AND @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
@@ -551,7 +551,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[6], v[0]},
 				{v[6], v[1]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[4]},
 				{v[0], v[1]},
@@ -564,7 +564,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[5], v[5]},
 				{v[5], v[4]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[5], v[0]},
 				{v[5], v[1]},
@@ -585,7 +585,7 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.LeftAntiJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[1], v[2]},
 				{v[1], v[3]},
@@ -593,12 +593,12 @@ func TestMergeJoiner(t *testing.T) {
 				{v[3], v[4]},
 				{v[3], v[5]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[2], v[2]},
 				{v[3], v[3]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[1], v[2]},
 				{v[1], v[3]},
@@ -618,18 +618,18 @@ func TestMergeJoiner(t *testing.T) {
 				Type: descpb.LeftAntiJoin,
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, v[2]},
 				{v[2], v[3]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{null, v[3]},
 				{v[2], v[4]},
 				{v[2], v[5]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{null, v[2]},
 			},
@@ -650,7 +650,7 @@ func TestMergeJoiner(t *testing.T) {
 				// Implicit AND @1 = @3 constraint.
 			},
 			outCols:   []uint32{0, 1},
-			leftTypes: randgen.TwoIntCols,
+			leftTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
@@ -661,7 +661,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[6], v[0]},
 				{v[6], v[1]},
 			},
-			rightTypes: randgen.TwoIntCols,
+			rightTypes: types.TwoIntCols,
 			rightInput: rowenc.EncDatumRows{
 				{v[0], v[4]},
 				{v[0], v[1]},
@@ -674,7 +674,7 @@ func TestMergeJoiner(t *testing.T) {
 				{v[5], v[5]},
 				{v[5], v[4]},
 			},
-			expectedTypes: randgen.TwoIntCols,
+			expectedTypes: types.TwoIntCols,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
@@ -778,8 +778,8 @@ func TestConsumerClosed(t *testing.T) {
 		// Implicit @1 = @2 constraint.
 	}
 	outCols := []uint32{0}
-	leftTypes := randgen.OneIntCol
-	rightTypes := randgen.OneIntCol
+	leftTypes := types.OneIntCol
+	rightTypes := types.OneIntCol
 
 	testCases := []struct {
 		typ       descpb.JoinType
@@ -876,8 +876,8 @@ func BenchmarkMergeJoiner(b *testing.B) {
 	for _, inputSize := range []int{0, 1 << 2, 1 << 4, 1 << 8, 1 << 12, 1 << 16} {
 		b.Run(fmt.Sprintf("InputSize=%d", inputSize), func(b *testing.B) {
 			rows := randgen.MakeIntRows(inputSize, numCols)
-			leftInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, rows)
-			rightInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, rows)
+			leftInput := execinfra.NewRepeatableRowSource(types.OneIntCol, rows)
+			rightInput := execinfra.NewRepeatableRowSource(types.OneIntCol, rows)
 			b.SetBytes(int64(8 * inputSize * numCols * 2))
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
@@ -895,8 +895,8 @@ func BenchmarkMergeJoiner(b *testing.B) {
 	for _, inputSize := range []int{0, 1 << 2, 1 << 4, 1 << 8, 1 << 12, 1 << 16} {
 		numRepeats := inputSize
 		b.Run(fmt.Sprintf("OneSideRepeatInputSize=%d", inputSize), func(b *testing.B) {
-			leftInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, randgen.MakeIntRows(inputSize, numCols))
-			rightInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, randgen.MakeRepeatedIntRows(numRepeats, inputSize, numCols))
+			leftInput := execinfra.NewRepeatableRowSource(types.OneIntCol, randgen.MakeIntRows(inputSize, numCols))
+			rightInput := execinfra.NewRepeatableRowSource(types.OneIntCol, randgen.MakeRepeatedIntRows(numRepeats, inputSize, numCols))
 			b.SetBytes(int64(8 * inputSize * numCols * 2))
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
@@ -915,8 +915,8 @@ func BenchmarkMergeJoiner(b *testing.B) {
 		numRepeats := int(math.Sqrt(float64(inputSize)))
 		b.Run(fmt.Sprintf("BothSidesRepeatInputSize=%d", inputSize), func(b *testing.B) {
 			row := randgen.MakeRepeatedIntRows(numRepeats, inputSize, numCols)
-			leftInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, row)
-			rightInput := execinfra.NewRepeatableRowSource(randgen.OneIntCol, row)
+			leftInput := execinfra.NewRepeatableRowSource(types.OneIntCol, row)
+			rightInput := execinfra.NewRepeatableRowSource(types.OneIntCol, row)
 			b.SetBytes(int64(8 * inputSize * numCols * 2))
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
