+++ pkg/sql/rowexec/project_set_test.go
@@ -48,13 +48,13 @@ func TestProjectSet(t *testing.T) {
 				Exprs: []execinfrapb.Expression{
 					{Expr: "@1 + 1"},
 				},
-				GeneratedColumns: randgen.OneIntCol,
+				GeneratedColumns: types.OneIntCol,
 				NumColsPerGen:    []uint32{1},
 			},
 			input: rowenc.EncDatumRows{
 				{v[2]},
 			},
-			inputTypes: randgen.OneIntCol,
+			inputTypes: types.OneIntCol,
 			expected: rowenc.EncDatumRows{
 				{v[2], v[3]},
 			},
@@ -65,14 +65,14 @@ func TestProjectSet(t *testing.T) {
 				Exprs: []execinfrapb.Expression{
 					{Expr: "generate_series(@1, 2)"},
 				},
-				GeneratedColumns: randgen.OneIntCol,
+				GeneratedColumns: types.OneIntCol,
 				NumColsPerGen:    []uint32{1},
 			},
 			input: rowenc.EncDatumRows{
 				{v[0]},
 				{v[1]},
 			},
-			inputTypes: randgen.OneIntCol,
+			inputTypes: types.OneIntCol,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0]},
 				{v[0], v[1]},
@@ -96,7 +96,7 @@ func TestProjectSet(t *testing.T) {
 			input: rowenc.EncDatumRows{
 				{v[0]},
 			},
-			inputTypes: randgen.OneIntCol,
+			inputTypes: types.OneIntCol,
 			expected: rowenc.EncDatumRows{
 				{v[0], v[0], v[0], v[0], v[0]},
 				{v[0], null, null, v[1], v[1]},
@@ -146,13 +146,13 @@ func BenchmarkProjectSet(b *testing.B) {
 				Exprs: []execinfrapb.Expression{
 					{Expr: "generate_series(1, 100000)"},
 				},
-				GeneratedColumns: randgen.OneIntCol,
+				GeneratedColumns: types.OneIntCol,
 				NumColsPerGen:    []uint32{1},
 			},
 			input: rowenc.EncDatumRows{
 				{v[0]},
 			},
-			inputTypes: randgen.OneIntCol,
+			inputTypes: types.OneIntCol,
 		},
 	}
