+++ pkg/sql/rowexec/set_op_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
-	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
@@ -91,7 +90,7 @@ func intersectAllTestCases() []setOpTestCase {
 			// Check that INTERSECT ALL only returns rows that are in both the left
 			// and right side.
 			setOpType:   descpb.IntersectAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, null},
@@ -136,7 +135,7 @@ func intersectAllTestCases() []setOpTestCase {
 			// Check that INTERSECT ALL returns the correct number of duplicates when
 			// the left side contains more duplicates of a row than the right side.
 			setOpType:   descpb.IntersectAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, null},
@@ -174,7 +173,7 @@ func intersectAllTestCases() []setOpTestCase {
 			// Check that INTERSECT ALL returns the correct number of duplicates when
 			// the right side contains more duplicates of a row than the left side.
 			setOpType:   descpb.IntersectAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, v[0]},
@@ -224,7 +223,7 @@ func exceptAllTestCases() []setOpTestCase {
 			// Check that EXCEPT ALL only returns rows that are on the left side
 			// but not the right side.
 			setOpType:   descpb.ExceptAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, null},
@@ -264,7 +263,7 @@ func exceptAllTestCases() []setOpTestCase {
 			// Check that EXCEPT ALL returns the correct number of duplicates when
 			// the left side contains more duplicates of a row than the right side.
 			setOpType:   descpb.ExceptAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, null},
@@ -300,7 +299,7 @@ func exceptAllTestCases() []setOpTestCase {
 			// Check that EXCEPT ALL returns the correct number of duplicates when
 			// the right side contains more duplicates of a row than the left side.
 			setOpType:   descpb.ExceptAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			leftInput: rowenc.EncDatumRows{
 				{null, null},
 				{null, v[0]},
@@ -332,7 +331,7 @@ func exceptAllTestCases() []setOpTestCase {
 		{
 			// Check that EXCEPT ALL handles mixed ordering correctly.
 			setOpType:   descpb.ExceptAllJoin,
-			columnTypes: randgen.TwoIntCols,
+			columnTypes: types.TwoIntCols,
 			ordering: colinfo.ColumnOrdering{
 				{ColIdx: 0, Direction: encoding.Descending},
 				{ColIdx: 1, Direction: encoding.Ascending},
