+++ pkg/sql/rowexec/sorter_test.go
@@ -66,7 +66,7 @@ func TestSorter(t *testing.T) {
 						{ColIdx: 2, Direction: asc},
 					}),
 			},
-			types: randgen.ThreeIntCols,
+			types: types.ThreeIntCols,
 			input: rowenc.EncDatumRows{
 				{v[1], v[0], v[4]},
 				{v[3], v[4], v[1]},
@@ -97,7 +97,7 @@ func TestSorter(t *testing.T) {
 					}),
 			},
 			post:  execinfrapb.PostProcessSpec{Limit: 4},
-			types: randgen.ThreeIntCols,
+			types: types.ThreeIntCols,
 			input: rowenc.EncDatumRows{
 				{v[3], v[3], v[0]},
 				{v[3], v[4], v[1]},
@@ -125,7 +125,7 @@ func TestSorter(t *testing.T) {
 					}),
 			},
 			post:  execinfrapb.PostProcessSpec{Offset: 2, Limit: 2},
-			types: randgen.ThreeIntCols,
+			types: types.ThreeIntCols,
 			input: rowenc.EncDatumRows{
 				{v[3], v[3], v[0]},
 				{v[3], v[4], v[1]},
@@ -151,7 +151,7 @@ func TestSorter(t *testing.T) {
 						{ColIdx: 2, Direction: asc},
 					}),
 			},
-			types: randgen.ThreeIntCols,
+			types: types.ThreeIntCols,
 			input: rowenc.EncDatumRows{
 				{v[0], v[1], v[2]},
 				{v[0], v[1], v[0]},
@@ -427,7 +427,7 @@ func BenchmarkSortAll(b *testing.B) {
 
 	for _, numRows := range []int{1 << 4, 1 << 8, 1 << 12, 1 << 16} {
 		b.Run(fmt.Sprintf("rows=%d", numRows), func(b *testing.B) {
-			input := execinfra.NewRepeatableRowSource(randgen.TwoIntCols, randgen.MakeRandIntRows(rng, numRows, numCols))
+			input := execinfra.NewRepeatableRowSource(types.TwoIntCols, randgen.MakeRandIntRows(rng, numRows, numCols))
 			b.SetBytes(int64(numRows * numCols * 8))
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
@@ -469,7 +469,7 @@ func BenchmarkSortLimit(b *testing.B) {
 
 	const numRows = 1 << 16
 	b.Run(fmt.Sprintf("rows=%d", numRows), func(b *testing.B) {
-		input := execinfra.NewRepeatableRowSource(randgen.TwoIntCols, randgen.MakeRandIntRows(rng, numRows, numCols))
+		input := execinfra.NewRepeatableRowSource(types.TwoIntCols, randgen.MakeRandIntRows(rng, numRows, numCols))
 		for _, limit := range []uint64{1 << 4, 1 << 8, 1 << 12, 1 << 16} {
 			post := execinfrapb.PostProcessSpec{Limit: limit}
 			b.Run(fmt.Sprintf("Limit=%d", limit), func(b *testing.B) {
@@ -525,7 +525,7 @@ func BenchmarkSortChunks(b *testing.B) {
 				for i, row := range rows {
 					row[0] = randgen.IntEncDatum(i / chunkSize)
 				}
-				input := execinfra.NewRepeatableRowSource(randgen.TwoIntCols, rows)
+				input := execinfra.NewRepeatableRowSource(types.TwoIntCols, rows)
 				b.SetBytes(int64(numRows * numCols * 8))
 				b.ResetTimer()
 				for i := 0; i < b.N; i++ {
