+++ pkg/sql/types/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     srcs = [
         "alias.go",
         "oid.go",
+        "testutils.go",
         "types.go",
         "types_jsonpb.go",
     ],
